requirements.txt
Django~=5.0.7
ipython==8.26.0
psycopg2==2.9.9
jedi==0.19.1
git+https://github.com/dbca-wa/ledger_api_client.git
django-extensions~=3.2.3
git+https://github.com/dbca-wa/dpaw-utils.git@0.4.2#egg=dpaw-utils
coverage~=7.6.1
coveralls~=4.0.1
reportlab~=4.2.2
django_bootstrap3~=24.2
django-braces>=1.8.1
django-datatables-view~=1.2.0
django-reversion~=5.0.12
django-reversion-compare~=0.17.0
git+https://github.com/xzzy/django-preserialize.git#egg=django-preserialize
django-countries~=7.6.1
django-cron==0.6.0
django-dynamic-fixture~=4.0.1
openpyxl~=3.1.5
datapackage~=1.15.4
python-dateutil~=2.9.0
py4j~=0.10.9.7
djangorestframework==3.15.2
djangorestframework-csv~=3.0.2
pycountry~=24.6.1
six>=1.10.0
django-utils-six==2.0
#django-ical==1.4
#django-taggit==0.21.3
geojson~=3.1.0
unicodecsv==0.14.1
PyPDF2~=3.0.1
django-extra-fields~=3.0.2
django-crispy-forms~=2.3
django-reset-migrations==0.4.0
djangorestframework-datatables~=0.7.0
django-confy==1.0.4
xlwt==1.3.0
xlsxwriter~=3.2.0
gunicorn==23.0.0
whitenoise==6.7.0
mixer~=7.2.2
django-dirtyfields==1.9.3
python-docx~=1.1.2
docxtpl~=0.10.0
dj-database-url==2.2.0
#psutil==6.0.0
phonenumbers~=8.13.42
#django-treebeard==4.3.1
#django-multiselectfield==0.1.12
#django-smart-selects==1.5.9
git+https://github.com/jazzband/django-smart-selects.git@1.7.1
pandas~=2.2.3
git+https://github.com/dbca-wa/django-media-serv.git#egg=django_media_serv
django-widget-tweaks~=1.5.0
django-phonenumber-field==8.0.0
babel~=2.15.0
git+https://github.com/dbca-wa/webtemplate2.git#egg=webtemplate_dbca
tqdm~=4.66.5
python-decouple==3.8
git+https://github.com/dbca-wa/appmonitor_client.git#egg=appmonitor_client
click==8.1.7
numpy~=2.1.1
setuptools~=72.1.0
wheel==0.44.0
git+https://github.com/dbca-wa/django5-utils-six.git#egg=django5_six
crispy_bootstrap5==2024.2